requirements.txt
# PGG情感记忆生成系统 - 依赖包
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
numpy==1.24.3
librosa==0.10.1
soundfile==0.12.1
Pillow==10.0.1
pymongo==4.5.0
python-speech-features==0.6
scipy==1.11.3
pyaudio==0.2.11
wave==0.0.2
json5==0.9.14
python-dotenv==1.0.0
psutil==5.9.5
# 预留AI模型相关依赖
torch==2.0.1
transformers==4.33.2
# diffusers==0.21.4  # 用于Stable Diffusion
# openai==0.28.1  # 用于OpenAI API调用
pandas==2.0.3 